requirements.txt
pdfplumber>=0.11.0
striprtf>=0.0.26
python-dotenv>=1.0.1
certifi>=2024.8.30
playwright>=1.49.0
tf-playwright-stealth>=1.1.0
